SessionUTs: Adding coverage
SessionUTs: Adding coverage for the new parameter. Splitting up test_session.py uts into two files. Make them smaller and easier to maintain.
